Build guide isnt maintained and incomplete
Build guide hasnt been up to date for a long time. A lot has changed since then. Especially the stance on building your own MMC install. This should be revised, once debrand is done and a new guide is made. For now, better remove it, then suggesting to build MMC for yourself.